Conflicts:
	src/external/tng_io/src/lib/tng_io.c

+++ src/gromacs/gmxpreprocess/topio.c
@@ -995,7 +995,7 @@ static char **read_topol(const char *infile, const char *outfile,
                             break;
                         }
                         default:
-                            fprintf (stderr, "case: %d\n", d);
+                            fprintf (stderr, "case: %d\n", (int)d);
                             gmx_incons("unknown directive");
                     }
                 }
